[dbs-leipzig#1571] Code cleanup and changed all result values to float

+++ gradoop-temporal/src/main/java/org/gradoop/temporal/model/impl/operators/metric/DegreeVarianceEvolution.java
@@ -16,13 +16,16 @@
 package org.gradoop.temporal.model.impl.operators.metric;
 
 import org.apache.flink.api.java.DataSet;
-import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.api.java.tuple.Tuple3;
 import org.gradoop.flink.model.api.operators.UnaryBaseGraphToValueOperator;
 import org.gradoop.flink.model.impl.operators.sampling.functions.VertexDegree;
 import org.gradoop.temporal.model.api.TimeDimension;
 import org.gradoop.temporal.model.impl.TemporalGraph;
-import org.gradoop.temporal.model.impl.operators.metric.functions.*;
+import org.gradoop.temporal.model.impl.operators.metric.functions.BuildTemporalDegreeTree;
+import org.gradoop.temporal.model.impl.operators.metric.functions.FlatMapVertexIdEdgeInterval;
+import org.gradoop.temporal.model.impl.operators.metric.functions.GroupDegreeTreesToVariance;
+import org.gradoop.temporal.model.impl.operators.metric.functions.TransformDeltaToAbsoluteDegreeTree;
+import org.gradoop.temporal.model.impl.operators.metric.functions.MapDegreesToInterval;
 
 import java.util.Objects;
 
@@ -30,41 +33,41 @@ import java.util.Objects;
  * Operator that calculates the degree variance evolution of a temporal graph for the
  * whole lifetime of the graph.
  */
-public class DegreeVarianceEvolution implements UnaryBaseGraphToValueOperator<TemporalGraph, DataSet<Tuple3<Long, Long, Double>>> {
-    /**
-     * The time dimension that will be considered.
-     */
-    private final TimeDimension dimension;
+public class DegreeVarianceEvolution implements UnaryBaseGraphToValueOperator<TemporalGraph, DataSet<Tuple3<Long, Long, Float>>> {
+  /**
+   * The time dimension that will be considered.
+   */
+  private final TimeDimension dimension;
 
-    /**
-     * The degree type (IN, OUT, BOTH);
-     */
-    private final VertexDegree degreeType;
+  /**
+   * The degree type (IN, OUT, BOTH);
+   */
+  private final VertexDegree degreeType;
 
-    /**
-     * Creates an instance of this average degree evolution operator.
-     *
-     * @param degreeType the degree type to use (IN, OUT, BOTH).
-     * @param dimension  the time dimension to use (VALID_TIME, TRANSACTION_TIME).
-     */
-    public DegreeVarianceEvolution(VertexDegree degreeType, TimeDimension dimension) {
-        this.degreeType = Objects.requireNonNull(degreeType);
-        this.dimension = Objects.requireNonNull(dimension);
-    }
+  /**
+   * Creates an instance of this average degree evolution operator.
+   *
+   * @param degreeType the degree type to use (IN, OUT, BOTH).
+   * @param dimension  the time dimension to use (VALID_TIME, TRANSACTION_TIME).
+   */
+  public DegreeVarianceEvolution(VertexDegree degreeType, TimeDimension dimension) {
+    this.degreeType = Objects.requireNonNull(degreeType);
+    this.dimension = Objects.requireNonNull(dimension);
+  }
 
-    @Override
-    public DataSet<Tuple3<Long, Long, Double>> execute(TemporalGraph graph) {
-        return graph.getEdges()
-                // 1) Extract vertex id(s) and corresponding time intervals
-                .flatMap(new FlatMapVertexIdEdgeInterval(dimension, degreeType))
-                // 2) Group them by the vertex id
-                .groupBy(0)
-                // 3) For each vertex id, build a degree tree data structure
-                .reduceGroup(new BuildTemporalDegreeTree())
-                // 4) Transform each tree to aggregated evolution
-                .map(new TransformDeltaToAbsoluteDegreeTree())
-                // 6) Merge trees together and calculate aggregation
-                .reduceGroup(new GroupDegreeTreesToVariance())
-                .mapPartition(new MapDegreesToInterval());
-    }
+  @Override
+  public DataSet<Tuple3<Long, Long, Float>> execute(TemporalGraph graph) {
+    return graph.getEdges()
+            // 1) Extract vertex id(s) and corresponding time intervals
+            .flatMap(new FlatMapVertexIdEdgeInterval(dimension, degreeType))
+            // 2) Group them by the vertex id
+            .groupBy(0)
+            // 3) For each vertex id, build a degree tree data structure
+            .reduceGroup(new BuildTemporalDegreeTree())
+            // 4) Transform each tree to aggregated evolution
+            .map(new TransformDeltaToAbsoluteDegreeTree())
+            // 6) Merge trees together and calculate aggregation
+            .reduceGroup(new GroupDegreeTreesToVariance())
+            .mapPartition(new MapDegreesToInterval());
+  }
 }
